@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { connect } from 'react-redux';
 import { Link, RouteComponentProps } from 'react-router-dom';
 import { Button, Col, Row, Table } from 'reactstrap';
@@ -12,90 +12,88 @@ import { APP_DATE_FORMAT, APP_LOCAL_DATE_FORMAT } from 'app/config/constants';
 
 export interface IBankAccountProps extends StateProps, DispatchProps, RouteComponentProps<{ url: string }> {}
 
-export class BankAccount extends React.Component<IBankAccountProps> {
-  componentDidMount() {
-    this.props.getEntities();
-  }
+export const BankAccount = (props: IBankAccountProps) => {
+  useEffect(() => {
+    props.getEntities();
+  }, []);
 
-  render() {
-    const { bankAccountList, match } = this.props;
-    return (
-      <div>
-        <h2 id="bank-account-heading">
-          <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.home.title">Bank Accounts</Translate>
-          <Link to={`${match.url}/new`} className="btn btn-primary float-right jh-create-entity" id="jh-create-entity">
-            <FontAwesomeIcon icon="plus" />
-            &nbsp;
-            <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.home.createLabel">Create a new Bank Account</Translate>
-          </Link>
-        </h2>
-        <div className="table-responsive">
-          {bankAccountList && bankAccountList.length > 0 ? (
-            <Table responsive aria-describedby="bank-account-heading">
-              <thead>
-                <tr>
-                  <th>
-                    <Translate contentKey="global.field.id">ID</Translate>
-                  </th>
-                  <th>
-                    <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.name">Name</Translate>
-                  </th>
-                  <th>
-                    <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.balance">Balance</Translate>
-                  </th>
-                  <th>
-                    <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.user">User</Translate>
-                  </th>
-                  <th />
-                </tr>
-              </thead>
-              <tbody>
-                {bankAccountList.map((bankAccount, i) => (
-                  <tr key={`entity-${i}`}>
-                    <td>
-                      <Button tag={Link} to={`${match.url}/${bankAccount.id}`} color="link" size="sm">
-                        {bankAccount.id}
+  const { bankAccountList, match } = props;
+  return (
+    <div>
+      <h2 id="bank-account-heading">
+        <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.home.title">Bank Accounts</Translate>
+        <Link to={`${match.url}/new`} className="btn btn-primary float-right jh-create-entity" id="jh-create-entity">
+          <FontAwesomeIcon icon="plus" />
+          &nbsp;
+          <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.home.createLabel">Create new Bank Account</Translate>
+        </Link>
+      </h2>
+      <div className="table-responsive">
+        {bankAccountList && bankAccountList.length > 0 ? (
+          <Table responsive>
+            <thead>
+              <tr>
+                <th>
+                  <Translate contentKey="global.field.id">ID</Translate>
+                </th>
+                <th>
+                  <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.name">Name</Translate>
+                </th>
+                <th>
+                  <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.balance">Balance</Translate>
+                </th>
+                <th>
+                  <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.user">User</Translate>
+                </th>
+                <th />
+              </tr>
+            </thead>
+            <tbody>
+              {bankAccountList.map((bankAccount, i) => (
+                <tr key={`entity-${i}`}>
+                  <td>
+                    <Button tag={Link} to={`${match.url}/${bankAccount.id}`} color="link" size="sm">
+                      {bankAccount.id}
+                    </Button>
+                  </td>
+                  <td>{bankAccount.name}</td>
+                  <td>{bankAccount.balance}</td>
+                  <td>{bankAccount.userLogin ? bankAccount.userLogin : ''}</td>
+                  <td className="text-right">
+                    <div className="btn-group flex-btn-group-container">
+                      <Button tag={Link} to={`${match.url}/${bankAccount.id}`} color="info" size="sm">
+                        <FontAwesomeIcon icon="eye" />{' '}
+                        <span className="d-none d-md-inline">
+                          <Translate contentKey="entity.action.view">View</Translate>
+                        </span>
                       </Button>
-                    </td>
-                    <td>{bankAccount.name}</td>
-                    <td>{bankAccount.balance}</td>
-                    <td>{bankAccount.userLogin ? bankAccount.userLogin : ''}</td>
-                    <td className="text-right">
-                      <div className="btn-group flex-btn-group-container">
-                        <Button tag={Link} to={`${match.url}/${bankAccount.id}`} color="info" size="sm">
-                          <FontAwesomeIcon icon="eye" />{' '}
-                          <span className="d-none d-md-inline">
-                            <Translate contentKey="entity.action.view">View</Translate>
-                          </span>
-                        </Button>
-                        <Button tag={Link} to={`${match.url}/${bankAccount.id}/edit`} color="primary" size="sm">
-                          <FontAwesomeIcon icon="pencil-alt" />{' '}
-                          <span className="d-none d-md-inline">
-                            <Translate contentKey="entity.action.edit">Edit</Translate>
-                          </span>
-                        </Button>
-                        <Button tag={Link} to={`${match.url}/${bankAccount.id}/delete`} color="danger" size="sm">
-                          <FontAwesomeIcon icon="trash" />{' '}
-                          <span className="d-none d-md-inline">
-                            <Translate contentKey="entity.action.delete">Delete</Translate>
-                          </span>
-                        </Button>
-                      </div>
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </Table>
-          ) : (
-            <div className="alert alert-warning">
-              <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.home.notFound">No Bank Accounts found</Translate>
-            </div>
-          )}
-        </div>
+                      <Button tag={Link} to={`${match.url}/${bankAccount.id}/edit`} color="primary" size="sm">
+                        <FontAwesomeIcon icon="pencil-alt" />{' '}
+                        <span className="d-none d-md-inline">
+                          <Translate contentKey="entity.action.edit">Edit</Translate>
+                        </span>
+                      </Button>
+                      <Button tag={Link} to={`${match.url}/${bankAccount.id}/delete`} color="danger" size="sm">
+                        <FontAwesomeIcon icon="trash" />{' '}
+                        <span className="d-none d-md-inline">
+                          <Translate contentKey="entity.action.delete">Delete</Translate>
+                        </span>
+                      </Button>
+                    </div>
+                  </td>
+                </tr>
+              ))}
+            </tbody>
+          </Table>
+        ) : (
+          <div className="alert alert-warning">
+            <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.home.notFound">No Bank Accounts found</Translate>
+          </div>
+        )}
       </div>
-    );
-  }
-}
+    </div>
+  );
+};
 
 const mapStateToProps = ({ bankAccount }: IRootState) => ({
   bankAccountList: bankAccount.entities
@@ -108,7 +106,4 @@ const mapDispatchToProps = {
 type StateProps = ReturnType<typeof mapStateToProps>;
 type DispatchProps = typeof mapDispatchToProps;
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(BankAccount);
+export default connect(mapStateToProps, mapDispatchToProps)(BankAccount);
