@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 import { RouteComponentProps } from 'react-router-dom';
 import { Modal, ModalHeader, ModalBody, ModalFooter, Button } from 'reactstrap';
@@ -11,52 +11,55 @@ import { getEntity, deleteEntity } from './label.reducer';
 
 export interface ILabelDeleteDialogProps extends StateProps, DispatchProps, RouteComponentProps<{ id: string }> {}
 
-export class LabelDeleteDialog extends React.Component<ILabelDeleteDialogProps> {
-  componentDidMount() {
-    this.props.getEntity(this.props.match.params.id);
-  }
+export const LabelDeleteDialog = (props: ILabelDeleteDialogProps) => {
+  useEffect(() => {
+    props.getEntity(props.match.params.id);
+  }, []);
 
-  confirmDelete = event => {
-    this.props.deleteEntity(this.props.labelEntity.id);
-    this.handleClose(event);
+  const handleClose = () => {
+    props.history.push('/label');
   };
 
-  handleClose = event => {
-    event.stopPropagation();
-    this.props.history.goBack();
+  useEffect(() => {
+    if (props.updateSuccess) {
+      handleClose();
+    }
+  }, [props.updateSuccess]);
+
+  const confirmDelete = () => {
+    props.deleteEntity(props.labelEntity.id);
   };
 
-  render() {
-    const { labelEntity } = this.props;
-    return (
-      <Modal isOpen toggle={this.handleClose}>
-        <ModalHeader toggle={this.handleClose}>
-          <Translate contentKey="entity.delete.title">Confirm delete operation</Translate>
-        </ModalHeader>
-        <ModalBody id="jhipsterSampleApplicationReactApp.label.delete.question">
-          <Translate contentKey="jhipsterSampleApplicationReactApp.label.delete.question" interpolate={{ id: labelEntity.id }}>
-            Are you sure you want to delete this Label?
-          </Translate>
-        </ModalBody>
-        <ModalFooter>
-          <Button color="secondary" onClick={this.handleClose}>
-            <FontAwesomeIcon icon="ban" />
-            &nbsp;
-            <Translate contentKey="entity.action.cancel">Cancel</Translate>
-          </Button>
-          <Button id="jhi-confirm-delete-label" color="danger" onClick={this.confirmDelete}>
-            <FontAwesomeIcon icon="trash" />
-            &nbsp;
-            <Translate contentKey="entity.action.delete">Delete</Translate>
-          </Button>
-        </ModalFooter>
-      </Modal>
-    );
-  }
-}
+  const { labelEntity } = props;
+  return (
+    <Modal isOpen toggle={handleClose}>
+      <ModalHeader toggle={handleClose}>
+        <Translate contentKey="entity.delete.title">Confirm delete operation</Translate>
+      </ModalHeader>
+      <ModalBody id="jhipsterSampleApplicationReactApp.label.delete.question">
+        <Translate contentKey="jhipsterSampleApplicationReactApp.label.delete.question" interpolate={{ id: labelEntity.id }}>
+          Are you sure you want to delete this Label?
+        </Translate>
+      </ModalBody>
+      <ModalFooter>
+        <Button color="secondary" onClick={handleClose}>
+          <FontAwesomeIcon icon="ban" />
+          &nbsp;
+          <Translate contentKey="entity.action.cancel">Cancel</Translate>
+        </Button>
+        <Button id="jhi-confirm-delete-label" color="danger" onClick={confirmDelete}>
+          <FontAwesomeIcon icon="trash" />
+          &nbsp;
+          <Translate contentKey="entity.action.delete">Delete</Translate>
+        </Button>
+      </ModalFooter>
+    </Modal>
+  );
+};
 
 const mapStateToProps = ({ label }: IRootState) => ({
-  labelEntity: label.entity
+  labelEntity: label.entity,
+  updateSuccess: label.updateSuccess
 });
 
 const mapDispatchToProps = { getEntity, deleteEntity };
@@ -64,7 +67,4 @@ const mapDispatchToProps = { getEntity, deleteEntity };
 type StateProps = ReturnType<typeof mapStateToProps>;
 type DispatchProps = typeof mapDispatchToProps;
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(LabelDeleteDialog);
+export default connect(mapStateToProps, mapDispatchToProps)(LabelDeleteDialog);
