@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 import { Link, RouteComponentProps } from 'react-router-dom';
 import { Button, Row, Col } from 'reactstrap';
@@ -12,45 +12,43 @@ import { APP_DATE_FORMAT, APP_LOCAL_DATE_FORMAT } from 'app/config/constants';
 
 export interface ILabelDetailProps extends StateProps, DispatchProps, RouteComponentProps<{ id: string }> {}
 
-export class LabelDetail extends React.Component<ILabelDetailProps> {
-  componentDidMount() {
-    this.props.getEntity(this.props.match.params.id);
-  }
+export const LabelDetail = (props: ILabelDetailProps) => {
+  useEffect(() => {
+    props.getEntity(props.match.params.id);
+  }, []);
 
-  render() {
-    const { labelEntity } = this.props;
-    return (
-      <Row>
-        <Col md="8">
-          <h2>
-            <Translate contentKey="jhipsterSampleApplicationReactApp.label.detail.title">Label</Translate> [<b>{labelEntity.id}</b>]
-          </h2>
-          <dl className="jh-entity-details">
-            <dt>
-              <span id="label">
-                <Translate contentKey="jhipsterSampleApplicationReactApp.label.label">Label</Translate>
-              </span>
-            </dt>
-            <dd>{labelEntity.label}</dd>
-          </dl>
-          <Button tag={Link} to="/label" replace color="info">
-            <FontAwesomeIcon icon="arrow-left" />{' '}
-            <span className="d-none d-md-inline">
-              <Translate contentKey="entity.action.back">Back</Translate>
+  const { labelEntity } = props;
+  return (
+    <Row>
+      <Col md="8">
+        <h2>
+          <Translate contentKey="jhipsterSampleApplicationReactApp.label.detail.title">Label</Translate> [<b>{labelEntity.id}</b>]
+        </h2>
+        <dl className="jh-entity-details">
+          <dt>
+            <span id="label">
+              <Translate contentKey="jhipsterSampleApplicationReactApp.label.label">Label</Translate>
             </span>
-          </Button>
-          &nbsp;
-          <Button tag={Link} to={`/label/${labelEntity.id}/edit`} replace color="primary">
-            <FontAwesomeIcon icon="pencil-alt" />{' '}
-            <span className="d-none d-md-inline">
-              <Translate contentKey="entity.action.edit">Edit</Translate>
-            </span>
-          </Button>
-        </Col>
-      </Row>
-    );
-  }
-}
+          </dt>
+          <dd>{labelEntity.label}</dd>
+        </dl>
+        <Button tag={Link} to="/label" replace color="info">
+          <FontAwesomeIcon icon="arrow-left" />{' '}
+          <span className="d-none d-md-inline">
+            <Translate contentKey="entity.action.back">Back</Translate>
+          </span>
+        </Button>
+        &nbsp;
+        <Button tag={Link} to={`/label/${labelEntity.id}/edit`} replace color="primary">
+          <FontAwesomeIcon icon="pencil-alt" />{' '}
+          <span className="d-none d-md-inline">
+            <Translate contentKey="entity.action.edit">Edit</Translate>
+          </span>
+        </Button>
+      </Col>
+    </Row>
+  );
+};
 
 const mapStateToProps = ({ label }: IRootState) => ({
   labelEntity: label.entity
@@ -61,7 +59,4 @@ const mapDispatchToProps = { getEntity };
 type StateProps = ReturnType<typeof mapStateToProps>;
 type DispatchProps = typeof mapDispatchToProps;
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(LabelDetail);
+export default connect(mapStateToProps, mapDispatchToProps)(LabelDetail);
