@@ -1,6 +1,7 @@
 package io.github.jhipster.sample;
 
-import io.github.jhipster.sample.config.DefaultProfileUtil;
+import io.github.jhipster.config.DefaultProfileUtil;
+
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 
