@@ -18,10 +18,7 @@ const defaultMiddlewares = [
 ];
 const composedMiddlewares = middlewares =>
   process.env.NODE_ENV === 'development'
-    ? compose(
-        applyMiddleware(...defaultMiddlewares, ...middlewares),
-        DevTools.instrument()
-      )
+    ? compose(applyMiddleware(...defaultMiddlewares, ...middlewares), DevTools.instrument())
     : compose(applyMiddleware(...defaultMiddlewares, ...middlewares));
 
 const initialize = (initialState?: IRootState, middlewares = []) => createStore(reducer, initialState, composedMiddlewares(middlewares));
