@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { connect } from 'react-redux';
 import { Link, RouteComponentProps } from 'react-router-dom';
 import { Button, Row, Col, Label } from 'reactstrap';
@@ -16,148 +16,134 @@ import { mapIdList } from 'app/shared/util/entity-utils';
 
 export interface IBankAccountUpdateProps extends StateProps, DispatchProps, RouteComponentProps<{ id: string }> {}
 
-export interface IBankAccountUpdateState {
-  isNew: boolean;
-  userId: string;
-}
+export const BankAccountUpdate = (props: IBankAccountUpdateProps) => {
+  const [userId, setUserId] = useState('0');
+  const [isNew, setIsNew] = useState(!props.match.params || !props.match.params.id);
 
-export class BankAccountUpdate extends React.Component<IBankAccountUpdateProps, IBankAccountUpdateState> {
-  constructor(props) {
-    super(props);
-    this.state = {
-      userId: '0',
-      isNew: !this.props.match.params || !this.props.match.params.id
-    };
-  }
+  const { bankAccountEntity, users, loading, updating } = props;
 
-  componentWillUpdate(nextProps, nextState) {
-    if (nextProps.updateSuccess !== this.props.updateSuccess && nextProps.updateSuccess) {
-      this.handleClose();
-    }
-  }
+  const handleClose = () => {
+    props.history.push('/bank-account');
+  };
 
-  componentDidMount() {
-    if (this.state.isNew) {
-      this.props.reset();
+  useEffect(() => {
+    if (isNew) {
+      props.reset();
     } else {
-      this.props.getEntity(this.props.match.params.id);
+      props.getEntity(props.match.params.id);
     }
 
-    this.props.getUsers();
-  }
+    props.getUsers();
+  }, []);
 
-  saveEntity = (event, errors, values) => {
+  useEffect(() => {
+    if (props.updateSuccess) {
+      handleClose();
+    }
+  }, [props.updateSuccess]);
+
+  const saveEntity = (event, errors, values) => {
     if (errors.length === 0) {
-      const { bankAccountEntity } = this.props;
       const entity = {
         ...bankAccountEntity,
         ...values
       };
 
-      if (this.state.isNew) {
-        this.props.createEntity(entity);
+      if (isNew) {
+        props.createEntity(entity);
       } else {
-        this.props.updateEntity(entity);
+        props.updateEntity(entity);
       }
     }
   };
 
-  handleClose = () => {
-    this.props.history.push('/bank-account');
-  };
-
-  render() {
-    const { bankAccountEntity, users, loading, updating } = this.props;
-    const { isNew } = this.state;
-
-    return (
-      <div>
-        <Row className="justify-content-center">
-          <Col md="8">
-            <h2 id="jhipsterSampleApplicationReactApp.bankAccount.home.createOrEditLabel">
-              <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.home.createOrEditLabel">
-                Create or edit a BankAccount
-              </Translate>
-            </h2>
-          </Col>
-        </Row>
-        <Row className="justify-content-center">
-          <Col md="8">
-            {loading ? (
-              <p>Loading...</p>
-            ) : (
-              <AvForm model={isNew ? {} : bankAccountEntity} onSubmit={this.saveEntity}>
-                {!isNew ? (
-                  <AvGroup>
-                    <Label for="bank-account-id">
-                      <Translate contentKey="global.field.id">ID</Translate>
-                    </Label>
-                    <AvInput id="bank-account-id" type="text" className="form-control" name="id" required readOnly />
-                  </AvGroup>
-                ) : null}
+  return (
+    <div>
+      <Row className="justify-content-center">
+        <Col md="8">
+          <h2 id="jhipsterSampleApplicationReactApp.bankAccount.home.createOrEditLabel">
+            <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.home.createOrEditLabel">
+              Create or edit a BankAccount
+            </Translate>
+          </h2>
+        </Col>
+      </Row>
+      <Row className="justify-content-center">
+        <Col md="8">
+          {loading ? (
+            <p>Loading...</p>
+          ) : (
+            <AvForm model={isNew ? {} : bankAccountEntity} onSubmit={saveEntity}>
+              {!isNew ? (
                 <AvGroup>
-                  <Label id="nameLabel" for="bank-account-name">
-                    <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.name">Name</Translate>
+                  <Label for="bank-account-id">
+                    <Translate contentKey="global.field.id">ID</Translate>
                   </Label>
-                  <AvField
-                    id="bank-account-name"
-                    type="text"
-                    name="name"
-                    validate={{
-                      required: { value: true, errorMessage: translate('entity.validation.required') }
-                    }}
-                  />
+                  <AvInput id="bank-account-id" type="text" className="form-control" name="id" required readOnly />
                 </AvGroup>
-                <AvGroup>
-                  <Label id="balanceLabel" for="bank-account-balance">
-                    <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.balance">Balance</Translate>
-                  </Label>
-                  <AvField
-                    id="bank-account-balance"
-                    type="text"
-                    name="balance"
-                    validate={{
-                      required: { value: true, errorMessage: translate('entity.validation.required') },
-                      number: { value: true, errorMessage: translate('entity.validation.number') }
-                    }}
-                  />
-                </AvGroup>
-                <AvGroup>
-                  <Label for="bank-account-user">
-                    <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.user">User</Translate>
-                  </Label>
-                  <AvInput id="bank-account-user" type="select" className="form-control" name="userId">
-                    <option value="" key="0" />
-                    {users
-                      ? users.map(otherEntity => (
-                          <option value={otherEntity.id} key={otherEntity.id}>
-                            {otherEntity.login}
-                          </option>
-                        ))
-                      : null}
-                  </AvInput>
-                </AvGroup>
-                <Button tag={Link} id="cancel-save" to="/bank-account" replace color="info">
-                  <FontAwesomeIcon icon="arrow-left" />
-                  &nbsp;
-                  <span className="d-none d-md-inline">
-                    <Translate contentKey="entity.action.back">Back</Translate>
-                  </span>
-                </Button>
+              ) : null}
+              <AvGroup>
+                <Label id="nameLabel" for="bank-account-name">
+                  <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.name">Name</Translate>
+                </Label>
+                <AvField
+                  id="bank-account-name"
+                  type="text"
+                  name="name"
+                  validate={{
+                    required: { value: true, errorMessage: translate('entity.validation.required') }
+                  }}
+                />
+              </AvGroup>
+              <AvGroup>
+                <Label id="balanceLabel" for="bank-account-balance">
+                  <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.balance">Balance</Translate>
+                </Label>
+                <AvField
+                  id="bank-account-balance"
+                  type="text"
+                  name="balance"
+                  validate={{
+                    required: { value: true, errorMessage: translate('entity.validation.required') },
+                    number: { value: true, errorMessage: translate('entity.validation.number') }
+                  }}
+                />
+              </AvGroup>
+              <AvGroup>
+                <Label for="bank-account-user">
+                  <Translate contentKey="jhipsterSampleApplicationReactApp.bankAccount.user">User</Translate>
+                </Label>
+                <AvInput id="bank-account-user" type="select" className="form-control" name="userId">
+                  <option value="" key="0" />
+                  {users
+                    ? users.map(otherEntity => (
+                        <option value={otherEntity.id} key={otherEntity.id}>
+                          {otherEntity.login}
+                        </option>
+                      ))
+                    : null}
+                </AvInput>
+              </AvGroup>
+              <Button tag={Link} id="cancel-save" to="/bank-account" replace color="info">
+                <FontAwesomeIcon icon="arrow-left" />
                 &nbsp;
-                <Button color="primary" id="save-entity" type="submit" disabled={updating}>
-                  <FontAwesomeIcon icon="save" />
-                  &nbsp;
-                  <Translate contentKey="entity.action.save">Save</Translate>
-                </Button>
-              </AvForm>
-            )}
-          </Col>
-        </Row>
-      </div>
-    );
-  }
-}
+                <span className="d-none d-md-inline">
+                  <Translate contentKey="entity.action.back">Back</Translate>
+                </span>
+              </Button>
+              &nbsp;
+              <Button color="primary" id="save-entity" type="submit" disabled={updating}>
+                <FontAwesomeIcon icon="save" />
+                &nbsp;
+                <Translate contentKey="entity.action.save">Save</Translate>
+              </Button>
+            </AvForm>
+          )}
+        </Col>
+      </Row>
+    </div>
+  );
+};
 
 const mapStateToProps = (storeState: IRootState) => ({
   users: storeState.userManagement.users,
@@ -178,7 +164,4 @@ const mapDispatchToProps = {
 type StateProps = ReturnType<typeof mapStateToProps>;
 type DispatchProps = typeof mapDispatchToProps;
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(BankAccountUpdate);
+export default connect(mapStateToProps, mapDispatchToProps)(BankAccountUpdate);
