@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import InfiniteScroll from 'react-infinite-scroller';
 import { connect } from 'react-redux';
 import { Link, RouteComponentProps } from 'react-router-dom';
@@ -14,157 +14,165 @@ import { ITEMS_PER_PAGE } from 'app/shared/util/pagination.constants';
 
 export interface IOperationProps extends StateProps, DispatchProps, RouteComponentProps<{ url: string }> {}
 
-export type IOperationState = IPaginationBaseState;
+export const Operation = (props: IOperationProps) => {
+  const [paginationState, setPaginationState] = useState(getSortState(props.location, ITEMS_PER_PAGE));
+  const [sorting, setSorting] = useState(false);
 
-export class Operation extends React.Component<IOperationProps, IOperationState> {
-  state: IOperationState = {
-    ...getSortState(this.props.location, ITEMS_PER_PAGE)
+  const getAllEntities = () => {
+    props.getEntities(paginationState.activePage - 1, paginationState.itemsPerPage, `${paginationState.sort},${paginationState.order}`);
   };
 
-  componentDidMount() {
-    this.reset();
-  }
-
-  componentDidUpdate() {
-    if (this.props.updateSuccess) {
-      this.reset();
-    }
-  }
-
-  reset = () => {
-    this.props.reset();
-    this.setState({ activePage: 1 }, () => {
-      this.getEntities();
+  const resetAll = () => {
+    props.reset();
+    setPaginationState({
+      ...paginationState,
+      activePage: 1
     });
   };
 
-  handleLoadMore = () => {
+  useEffect(() => {
+    resetAll();
+  }, []);
+
+  useEffect(() => {
+    if (props.updateSuccess) {
+      getAllEntities();
+    }
+  });
+
+  useEffect(() => {
+    getAllEntities();
+  }, [paginationState.activePage]);
+
+  const handleLoadMore = () => {
     if (window.pageYOffset > 0) {
-      this.setState({ activePage: this.state.activePage + 1 }, () => this.getEntities());
+      setPaginationState({
+        ...paginationState,
+        activePage: paginationState.activePage + 1
+      });
     }
   };
 
-  sort = prop => () => {
-    this.setState(
-      {
-        order: this.state.order === 'asc' ? 'desc' : 'asc',
-        sort: prop
-      },
-      () => {
-        this.reset();
-      }
-    );
+  useEffect(() => {
+    if (sorting) {
+      getAllEntities();
+      setSorting(false);
+    }
+  }, [sorting]);
+
+  const sort = p => () => {
+    props.reset();
+    setPaginationState({
+      ...paginationState,
+      activePage: 1,
+      order: paginationState.order === 'asc' ? 'desc' : 'asc',
+      sort: p
+    });
+    setSorting(true);
   };
 
-  getEntities = () => {
-    const { activePage, itemsPerPage, sort, order } = this.state;
-    this.props.getEntities(activePage - 1, itemsPerPage, `${sort},${order}`);
-  };
-
-  render() {
-    const { operationList, match } = this.props;
-    return (
-      <div>
-        <h2 id="operation-heading">
-          <Translate contentKey="jhipsterSampleApplicationReactApp.operation.home.title">Operations</Translate>
-          <Link to={`${match.url}/new`} className="btn btn-primary float-right jh-create-entity" id="jh-create-entity">
-            <FontAwesomeIcon icon="plus" />
-            &nbsp;
-            <Translate contentKey="jhipsterSampleApplicationReactApp.operation.home.createLabel">Create a new Operation</Translate>
-          </Link>
-        </h2>
-        <div className="table-responsive">
-          <InfiniteScroll
-            pageStart={this.state.activePage}
-            loadMore={this.handleLoadMore}
-            hasMore={this.state.activePage - 1 < this.props.links.next}
-            loader={<div className="loader">Loading ...</div>}
-            threshold={0}
-            initialLoad={false}
-          >
-            {operationList && operationList.length > 0 ? (
-              <Table responsive aria-describedby="operation-heading">
-                <thead>
-                  <tr>
-                    <th className="hand" onClick={this.sort('id')}>
-                      <Translate contentKey="global.field.id">ID</Translate> <FontAwesomeIcon icon="sort" />
-                    </th>
-                    <th className="hand" onClick={this.sort('date')}>
-                      <Translate contentKey="jhipsterSampleApplicationReactApp.operation.date">Date</Translate>{' '}
-                      <FontAwesomeIcon icon="sort" />
-                    </th>
-                    <th className="hand" onClick={this.sort('description')}>
-                      <Translate contentKey="jhipsterSampleApplicationReactApp.operation.description">Description</Translate>{' '}
-                      <FontAwesomeIcon icon="sort" />
-                    </th>
-                    <th className="hand" onClick={this.sort('amount')}>
-                      <Translate contentKey="jhipsterSampleApplicationReactApp.operation.amount">Amount</Translate>{' '}
-                      <FontAwesomeIcon icon="sort" />
-                    </th>
-                    <th>
-                      <Translate contentKey="jhipsterSampleApplicationReactApp.operation.bankAccount">Bank Account</Translate>{' '}
-                      <FontAwesomeIcon icon="sort" />
-                    </th>
-                    <th />
-                  </tr>
-                </thead>
-                <tbody>
-                  {operationList.map((operation, i) => (
-                    <tr key={`entity-${i}`}>
-                      <td>
-                        <Button tag={Link} to={`${match.url}/${operation.id}`} color="link" size="sm">
-                          {operation.id}
+  const { operationList, match } = props;
+  return (
+    <div>
+      <h2 id="operation-heading">
+        <Translate contentKey="jhipsterSampleApplicationReactApp.operation.home.title">Operations</Translate>
+        <Link to={`${match.url}/new`} className="btn btn-primary float-right jh-create-entity" id="jh-create-entity">
+          <FontAwesomeIcon icon="plus" />
+          &nbsp;
+          <Translate contentKey="jhipsterSampleApplicationReactApp.operation.home.createLabel">Create new Operation</Translate>
+        </Link>
+      </h2>
+      <div className="table-responsive">
+        <InfiniteScroll
+          pageStart={paginationState.activePage}
+          loadMore={handleLoadMore}
+          hasMore={paginationState.activePage - 1 < props.links.next}
+          loader={<div className="loader">Loading ...</div>}
+          threshold={0}
+          initialLoad={false}
+        >
+          {operationList && operationList.length > 0 ? (
+            <Table responsive>
+              <thead>
+                <tr>
+                  <th className="hand" onClick={sort('id')}>
+                    <Translate contentKey="global.field.id">ID</Translate> <FontAwesomeIcon icon="sort" />
+                  </th>
+                  <th className="hand" onClick={sort('date')}>
+                    <Translate contentKey="jhipsterSampleApplicationReactApp.operation.date">Date</Translate>{' '}
+                    <FontAwesomeIcon icon="sort" />
+                  </th>
+                  <th className="hand" onClick={sort('description')}>
+                    <Translate contentKey="jhipsterSampleApplicationReactApp.operation.description">Description</Translate>{' '}
+                    <FontAwesomeIcon icon="sort" />
+                  </th>
+                  <th className="hand" onClick={sort('amount')}>
+                    <Translate contentKey="jhipsterSampleApplicationReactApp.operation.amount">Amount</Translate>{' '}
+                    <FontAwesomeIcon icon="sort" />
+                  </th>
+                  <th>
+                    <Translate contentKey="jhipsterSampleApplicationReactApp.operation.bankAccount">Bank Account</Translate>{' '}
+                    <FontAwesomeIcon icon="sort" />
+                  </th>
+                  <th />
+                </tr>
+              </thead>
+              <tbody>
+                {operationList.map((operation, i) => (
+                  <tr key={`entity-${i}`}>
+                    <td>
+                      <Button tag={Link} to={`${match.url}/${operation.id}`} color="link" size="sm">
+                        {operation.id}
+                      </Button>
+                    </td>
+                    <td>
+                      <TextFormat type="date" value={operation.date} format={APP_DATE_FORMAT} />
+                    </td>
+                    <td>{operation.description}</td>
+                    <td>{operation.amount}</td>
+                    <td>
+                      {operation.bankAccountName ? (
+                        <Link to={`bank-account/${operation.bankAccountId}`}>{operation.bankAccountName}</Link>
+                      ) : (
+                        ''
+                      )}
+                    </td>
+                    <td className="text-right">
+                      <div className="btn-group flex-btn-group-container">
+                        <Button tag={Link} to={`${match.url}/${operation.id}`} color="info" size="sm">
+                          <FontAwesomeIcon icon="eye" />{' '}
+                          <span className="d-none d-md-inline">
+                            <Translate contentKey="entity.action.view">View</Translate>
+                          </span>
                         </Button>
-                      </td>
-                      <td>
-                        <TextFormat type="date" value={operation.date} format={APP_DATE_FORMAT} />
-                      </td>
-                      <td>{operation.description}</td>
-                      <td>{operation.amount}</td>
-                      <td>
-                        {operation.bankAccountName ? (
-                          <Link to={`bank-account/${operation.bankAccountId}`}>{operation.bankAccountName}</Link>
-                        ) : (
-                          ''
-                        )}
-                      </td>
-                      <td className="text-right">
-                        <div className="btn-group flex-btn-group-container">
-                          <Button tag={Link} to={`${match.url}/${operation.id}`} color="info" size="sm">
-                            <FontAwesomeIcon icon="eye" />{' '}
-                            <span className="d-none d-md-inline">
-                              <Translate contentKey="entity.action.view">View</Translate>
-                            </span>
-                          </Button>
-                          <Button tag={Link} to={`${match.url}/${operation.id}/edit`} color="primary" size="sm">
-                            <FontAwesomeIcon icon="pencil-alt" />{' '}
-                            <span className="d-none d-md-inline">
-                              <Translate contentKey="entity.action.edit">Edit</Translate>
-                            </span>
-                          </Button>
-                          <Button tag={Link} to={`${match.url}/${operation.id}/delete`} color="danger" size="sm">
-                            <FontAwesomeIcon icon="trash" />{' '}
-                            <span className="d-none d-md-inline">
-                              <Translate contentKey="entity.action.delete">Delete</Translate>
-                            </span>
-                          </Button>
-                        </div>
-                      </td>
-                    </tr>
-                  ))}
-                </tbody>
-              </Table>
-            ) : (
-              <div className="alert alert-warning">
-                <Translate contentKey="jhipsterSampleApplicationReactApp.operation.home.notFound">No Operations found</Translate>
-              </div>
-            )}
-          </InfiniteScroll>
-        </div>
+                        <Button tag={Link} to={`${match.url}/${operation.id}/edit`} color="primary" size="sm">
+                          <FontAwesomeIcon icon="pencil-alt" />{' '}
+                          <span className="d-none d-md-inline">
+                            <Translate contentKey="entity.action.edit">Edit</Translate>
+                          </span>
+                        </Button>
+                        <Button tag={Link} to={`${match.url}/${operation.id}/delete`} color="danger" size="sm">
+                          <FontAwesomeIcon icon="trash" />{' '}
+                          <span className="d-none d-md-inline">
+                            <Translate contentKey="entity.action.delete">Delete</Translate>
+                          </span>
+                        </Button>
+                      </div>
+                    </td>
+                  </tr>
+                ))}
+              </tbody>
+            </Table>
+          ) : (
+            <div className="alert alert-warning">
+              <Translate contentKey="jhipsterSampleApplicationReactApp.operation.home.notFound">No Operations found</Translate>
+            </div>
+          )}
+        </InfiniteScroll>
       </div>
-    );
-  }
-}
+    </div>
+  );
+};
 
 const mapStateToProps = ({ operation }: IRootState) => ({
   operationList: operation.entities,
@@ -182,7 +190,4 @@ const mapDispatchToProps = {
 type StateProps = ReturnType<typeof mapStateToProps>;
 type DispatchProps = typeof mapDispatchToProps;
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(Operation);
+export default connect(mapStateToProps, mapDispatchToProps)(Operation);
