@@ -63,7 +63,4 @@ const mapDispatchToProps = { activateAction, reset };
 type StateProps = ReturnType<typeof mapStateToProps>;
 type DispatchProps = typeof mapDispatchToProps;
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(ActivatePage);
+export default connect(mapStateToProps, mapDispatchToProps)(ActivatePage);
