@@ -59,7 +59,4 @@ const mapDispatchToProps = { handlePasswordResetInit, reset };
 
 type DispatchProps = typeof mapDispatchToProps;
 
-export default connect(
-  null,
-  mapDispatchToProps
-)(PasswordResetInit);
+export default connect(null, mapDispatchToProps)(PasswordResetInit);
