@@ -95,7 +95,4 @@ const mapDispatchToProps = { systemHealth };
 type StateProps = ReturnType<typeof mapStateToProps>;
 type DispatchProps = typeof mapDispatchToProps;
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(HealthPage);
+export default connect(mapStateToProps, mapDispatchToProps)(HealthPage);
