@@ -130,7 +130,4 @@ const mapDispatchToProps = { getLoggers, changeLogLevel };
 type StateProps = ReturnType<typeof mapStateToProps>;
 type DispatchProps = typeof mapDispatchToProps;
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(LogsPage);
+export default connect(mapStateToProps, mapDispatchToProps)(LogsPage);
