@@ -128,7 +128,4 @@ const mapDispatchToProps = { systemMetrics, systemThreadDump };
 type StateProps = ReturnType<typeof mapStateToProps>;
 type DispatchProps = typeof mapDispatchToProps;
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(MetricsPage);
+export default connect(mapStateToProps, mapDispatchToProps)(MetricsPage);
