@@ -62,7 +62,4 @@ const mapDispatchToProps = { getUser, deleteUser };
 type StateProps = ReturnType<typeof mapStateToProps>;
 type DispatchProps = typeof mapDispatchToProps;
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(UserManagementDeleteDialog);
+export default connect(mapStateToProps, mapDispatchToProps)(UserManagementDeleteDialog);
