@@ -112,7 +112,4 @@ const mapDispatchToProps = { getUser };
 type StateProps = ReturnType<typeof mapStateToProps>;
 type DispatchProps = typeof mapDispatchToProps;
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(UserManagementDetail);
+export default connect(mapStateToProps, mapDispatchToProps)(UserManagementDetail);
